Revert "Move Env out of the way so it can be imported in source and destination modules without causing a circular import"
This reverts commit 36af18c24e4eebab92b0f73d46e337d14061b798.
